added common file for inline content download form test cases
