+++ casestudy-webinar-ebook/test_case_study.py
@@ -2,11 +2,9 @@ import time
 import logging
 import pytest
 from common.utils import resize_browser
-from common.asserts import assert_content_download_thank_you_page
-from common.forms import submit_content_download_form
+from common.content_download_form import assert_content_download_inline_form, assert_required_fields, assert_invalid_names, assert_bad_email, assert_non_business_email, assert_success_download_form
 
 logger = logging.getLogger(__name__)
-
 
 @pytest.fixture(scope='function')
 def browser(module_browser, base_url, request):
@@ -17,52 +15,24 @@ def browser(module_browser, base_url, request):
 
 @pytest.mark.parametrize('browser', [('desktop_1'), ('mobile_1')], indirect=True)
 def test_content_download_inline_form(browser):
-    download_form = browser.find(xpath="//form[contains(@id, 'content-download-form')]")
-    assert download_form.is_displayed(), "inline form is not present"
-    h3_head = browser.find(xpath="//h3[contains(@id, 'report-heading')]")
-    assert h3_head and h3_head.is_displayed(), "Form h3 heading is not displayed"
-    sub_text = browser.find(xpath="//p[contains(@id, 'report-subtext')]")
-    assert sub_text and sub_text.is_displayed(), "Form sub-heading is not displayed"
+    assert_content_download_inline_form(browser)
 
 @pytest.mark.parametrize('browser', [('desktop_1'), ('mobile_1')], indirect=True)
 def test_required_fields(browser):
-    submit_content_download_form(browser)
-    email_error = browser.find(xpath="//label[@for='content-download-form-email'][@class='error']")
-    firstname_error = browser.find(xpath="//label[@for='content-download-form-first-name'][@class='error']")
-    lastname_error = browser.find(xpath="//label[@for='content-download-form-last-name'][@class='error']")
-    company_size_error = browser.find(xpath="//form[contains(@id, 'content-download-form')]//label[@for='number_of_employees'][@class='error']")
-    consent_error = browser.find(xpath="//form[contains(@id, 'content-download-form')]//label[@for='gdpr_consent'][@class='error']")
-    assert email_error and email_error.is_displayed(), "No error displayed for missing email"
-    assert firstname_error and firstname_error.is_displayed(), "No error displayed for missing firstname"
-    assert lastname_error and lastname_error.is_displayed(), "No error displayed for missing lastname"
-    assert company_size_error and company_size_error.is_displayed(), "No error displayed for missing company size"
-    assert consent_error and consent_error.is_displayed(), "No error displayed for missing checkbox"
+    assert_required_fields(browser)
 
 @pytest.mark.parametrize('browser', [('desktop_1'), ('mobile_1')], indirect=True)
 def test_invalid_names(browser):
-    submit_content_download_form(browser, firstname='test1', lastname='test2')
-    firstname_error = browser.find(xpath="//label[@for='content-download-form-first-name'][@class='error']")
-    lastname_error = browser.find(xpath="//label[@for='content-download-form-last-name'][@class='error']")
-    assert firstname_error and firstname_error.is_displayed(), "No error displayed for invalid firstname"
-    assert lastname_error and lastname_error.is_displayed(), "No error displayed for invalid lastname"
+    assert_invalid_names(browser)
 
 @pytest.mark.parametrize('browser', [('desktop_1'), ('mobile_1')], indirect=True)
 def test_bad_email(browser):
-    submit_content_download_form(browser, email='test')
-    email_error = browser.find(xpath="//label[@for='content-download-form-email'][@class='error']")
-    assert email_error and email_error.is_displayed(), 'No error displayed for invalid email'
+    assert_bad_email(browser)
 
 @pytest.mark.parametrize('browser', [('desktop_1'), ('mobile_1')], indirect=True)
 def test_non_business_email(browser):
-    submit_content_download_form(browser, email='test@gmail.com', firstname='test', lastname='test', company_size='Under 5', agree=True)
-    time.sleep(5)
-    email_error = browser.find(xpath="//label[@for='content-download-form-email'][@class='error']")
-    assert email_error and email_error.is_displayed(), 'No error displayed for non business email'
+    assert_non_business_email(browser)
 
-@pytest.mark.parametrize('browser', [('desktop_1')], indirect=True)
+@pytest.mark.parametrize('browser', [('desktop_1'), ('mobile_1')], indirect=True)
 def test_success_download_form(browser):
-    submit_content_download_form(browser, email='test@fyle.in', firstname='test', lastname='test', company_size='Under 5', agree=True)
-    time.sleep(5)
-    assert_content_download_thank_you_page(browser,'Realtime visibility into T&E for Zivame','test@fyle.in','https://cdn2.hubspot.net/hubfs/3906991/Case%20Study%20/Fyle-Zivame-Case-Study.pdf')
-    browser.back()
-
+    assert_success_download_form(browser)
