+++ expense-report-template/test_template_expense_report.py
@@ -2,7 +2,7 @@ import time
 import logging
 import pytest
 from common.utils import resize_browser
-from common.forms import submit_download_excel_form
+from common.content_download_form import assert_download_for_excel_form_modal, assert_required_fields, assert_invalid_names, assert_bad_email, assert_non_business_email, assert_success_download_form
 
 logger = logging.getLogger(__name__)
 
@@ -15,66 +15,28 @@ def browser(module_browser, base_url, request):
     return module_browser
 
 @pytest.mark.parametrize('browser', [('desktop_1'), ('mobile_1')], indirect=True)
-def test_download_for_excel_form_modal(browser):
-    download_modal_cta = browser.find(xpath="//a[contains(@id, 'download-excel')]")
-    time.sleep(2)
-    download_modal_cta.click()
-    time.sleep(2)
-    report_modal = browser.find(xpath="//div[contains(@id, 'expense-report-modal')]")
-    assert report_modal.is_displayed(), "modal is not opened"
-    time.sleep(2)
-    h3_head = browser.find(xpath="//h3[contains(@id, 'report-heading')]")
-    assert h3_head and h3_head.is_displayed(), "Modal h3 heading is not displayed"
-    sub_text = browser.find(xpath="//p[contains(@id, 'report-subtext')]")
-    assert sub_text and sub_text.is_displayed(), "Modal sub-heading is not displayed"
-    close_button = browser.find(xpath="//div[contains(@id, 'expense-report-modal')]//button[contains(@class, 'close')]")
-    close_button.click()
-    time.sleep(2)
-    assert not report_modal.is_displayed(), "modal is not closed"
+def test_content_download_inline_form(browser):
+    assert_download_for_excel_form_modal(browser)
 
 @pytest.mark.parametrize('browser', [('desktop_1'), ('mobile_1')], indirect=True)
 def test_required_fields(browser):
-    submit_download_excel_form(browser)
-    email_error = browser.find(xpath="//label[@for='content-download-form-email'][@class='error']")
-    firstname_error = browser.find(xpath="//label[@for='content-download-form-first-name'][@class='error']")
-    lastname_error = browser.find(xpath="//label[@for='content-download-form-last-name'][@class='error']")
-    company_size_error = browser.find(xpath="//form[contains(@id, 'content-download-form')]//label[@for='number_of_employees'][@class='error']")
-    consent_error = browser.find(xpath="//form[contains(@id, 'content-download-form')]//label[@for='gdpr_consent'][@class='error']")
-    assert email_error and email_error.is_displayed(), "No error displayed for missing email"
-    assert firstname_error and firstname_error.is_displayed(), "No error displayed for missing firstname"
-    assert lastname_error and lastname_error.is_displayed(), "No error displayed for missing lastname"
-    assert company_size_error and company_size_error.is_displayed(), "No error displayed for missing company size"
-    assert consent_error and consent_error.is_displayed(), "No error displayed for missing checkbox"
+    assert_required_fields(browser)
 
 @pytest.mark.parametrize('browser', [('desktop_1'), ('mobile_1')], indirect=True)
 def test_invalid_names(browser):
-    submit_download_excel_form(browser, firstname='test1', lastname='test2')
-    firstname_error = browser.find(xpath="//label[@for='content-download-form-first-name'][@class='error']")
-    lastname_error = browser.find(xpath="//label[@for='content-download-form-last-name'][@class='error']")
-    assert firstname_error and firstname_error.is_displayed(), "No error displayed for invalid firstname"
-    assert lastname_error and lastname_error.is_displayed(), "No error displayed for invalid lastname"
+    assert_invalid_names(browser)
 
 @pytest.mark.parametrize('browser', [('desktop_1'), ('mobile_1')], indirect=True)
 def test_bad_email(browser):
-    submit_download_excel_form(browser, email='test')
-    email_error = browser.find(xpath="//label[@for='content-download-form-email'][@class='error']")
-    assert email_error and email_error.is_displayed(), 'No error displayed for invalid email'
+    assert_bad_email(browser)
 
 @pytest.mark.parametrize('browser', [('desktop_1'), ('mobile_1')], indirect=True)
 def test_non_business_email(browser):
-    submit_download_excel_form(browser, email='test@gmail.com', firstname='test', lastname='test', company_size='Under 5', agree=True)
-    time.sleep(5)
-    email_error = browser.find(xpath="//label[@for='content-download-form-email'][@class='error']")
-    assert email_error and email_error.is_displayed(), 'No error displayed for non business email'
+    assert_non_business_email(browser)
 
 @pytest.mark.parametrize('browser', [('desktop_1'), ('mobile_1')], indirect=True)
 def test_success_download_form(browser):
-    submit_download_excel_form(browser, email='test@fyle.in', firstname='test', lastname='test', company_size='Under 5', agree=True)
-    time.sleep(2)
-    last_downloaded_filename = browser.get_downLoadeded_filename()
-    time.sleep(4)
-    browser.close_windows()
-    assert last_downloaded_filename == 'https://cdn2.hubspot.net/hubfs/3906991/simple-expense-report-template.xlsx', "Downloaded file is not correct"
+    assert_success_download_form(browser)
 
 @pytest.mark.parametrize('browser', [('desktop_1'), ('mobile_1')], indirect=True)
 def test_redirection_of_h2_cards(browser):
